% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paw_function.R
\name{paw}
\alias{paw}
\title{It shows the power of any number Mathematically
'bs' Base
'pw' Power
'paw' function}
\usage{
paw(bs, pw)
}
\description{
It shows the power of any number Mathematically
'bs' Base
'pw' Power
'paw' function
}
